refactor: build share logic extracted
